% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_delta.R
\name{early_delta}
\alias{early_delta}
\title{Compute Delta for Multiple Columns}
\usage{
early_delta(data, min_t, max_t, size, name)
}
\arguments{
\item{data}{The input data frame containing the source_group and group columns.}

\item{min_t}{The minimum time for the window.}

\item{max_t}{The maximum time for the window.}

\item{size}{An integer vector representing the sizes of each group.}

\item{name}{A character vector containing the names of each group (respective to the \code{size} vector).}
}
\value{
A matrix representing the delta values for each column along with associated count data.
}
\description{
Computes the delta values for multiple columns based on the binomial mixing matrix.
}
