% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_scenario.R
\name{extract_scenario}
\alias{extract_scenario}
\title{Extract Scenario}
\usage{
extract_scenario(scenarios)
}
\arguments{
\item{scenarios}{A list of scenarios.}
}
\value{
A dataframe with the extracted scenario parameters.
}
\description{
This function takes a list of scenarios and extracts the relevant parameters from each scenario to create a dataframe.
}
\keyword{internal}
