% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups_utils.R
\name{generate_clusters}
\alias{generate_clusters}
\title{Generate clustered data}
\usage{
generate_clusters(
  n_groups = NULL,
  size = NULL,
  name = NULL,
  spread = 0.5,
  shape = "random"
)
}
\arguments{
\item{n_groups}{the number of groups to generate.}

\item{size}{the number of individuals per group to generate. Default is 10 per group.}

\item{name}{the names of the groups. Default is "Group 1", "Group 2", ..., "Group n_groups".}

\item{spread}{the standard deviation of the x and y coordinates around the group means. Default is 0.5.}

\item{shape}{the shape of the clusters.Can be "random", "square", or "round". Default is "random".}
}
\description{
Adds x/y coords to individuals based on their groups
using K-means clustering to generate group means.
}
\keyword{internal}
