% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_M.R
\name{generate_Mcol}
\alias{generate_Mcol}
\title{Generate Mcol Matrix}
\usage{
generate_Mcol(group_n, size, gamma = rep(1, group_n), name = NULL)
}
\arguments{
\item{group_n}{An integer specifying the number of groups.}

\item{size}{A numeric vector specifying the size of each group.}

\item{gamma}{A numeric vector specifying the gamma values for each group. Default is a vector of 1s.}

\item{name}{A character vector specifying the names of the groups. Default is NULL, in which case the groups are named using the first 'group_n' letters of the alphabet.}
}
\value{
Transmission Allocation Matrix, rows are 'from' groups and columns are 'to' groups.
}
\description{
Transmission Allocation Matrix. Each row represent 'from' groups and columns represent 'to' groups.
Conditional on an individual in group 'i' seeding infection, the probability of transmission to group 'j' is given by Mcol\link{i,j}. The matrix is normalised such that the sum of each row equals 1.
}
\keyword{internal}
