% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups_utils.R
\name{generate_new_cases}
\alias{generate_new_cases}
\title{Generate new cases at time t}
\usage{
generate_new_cases(data, t, sources)
}
\arguments{
\item{data}{The current linelist at time t}

\item{t}{The current iteration of the timestep}

\item{sources}{The \code{sources} object returned by sample_sources}
}
\value{
A dataframe / linelist of new cases
}
\description{
Generate new cases at time t
}
\keyword{internal}
