% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups_utils.R
\name{generate_sequences}
\alias{generate_sequences}
\title{Generate DNA sequences based on given data, genome length, and mutation rate.}
\usage{
generate_sequences(data, genome_length, mutation_rate)
}
\arguments{
\item{data}{A data frame containing information about each case, including the source of infection (if known). Output of \code{simulate_groups()$data}.}

\item{genome_length}{An integer specifying the length of the DNA genome.}

\item{mutation_rate}{A numeric value representing the mutation rate per nucleotide per infection cycle.}
}
\value{
A character vector containing the generated DNA sequences.
}
\description{
This function generates DNA sequences based on the provided data, genome length, and mutation rate.
It iterates through each case in the data, generating a sequence by either randomly sampling nucleotides or mutating the sequence from a known source.
}
\examples{
data <- data.frame(
  id = c("case1", "case2", "case3"),
  source = c(NA, "case1", "case2")
)
generate_sequences(data, genome_length = 1000, mutation_rate = 0.001)

}
