% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamma.R
\name{get_gamma}
\alias{get_gamma}
\title{Calculate Assortativity Coefficient (Gamma)}
\usage{
get_gamma(beta, f)
}
\arguments{
\item{beta}{Numeric, the observed within-group transmission probability.}

\item{f}{Numeric, the fraction of the population that is in the group.}
}
\value{
Numeric, the assortativity coefficient (Gamma) computed using the provided parameters.
}
\description{
The assortativity coefficient (Gamma) measures the degree to which individuals within a group
are more likely to interact with each other than with individuals from other groups.
}
\examples{
# Calculate the assortativity coefficient for a group
beta <- 0.6
f <- 0.4
gamma <- get_gamma(beta, f)
print(gamma)

}
