% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups_utils.R
\name{get_gt}
\alias{get_gt}
\title{Compute generation time values}
\usage{
get_gt(data)
}
\arguments{
\item{data}{The data element of \code{simulate_groups}.}
}
\value{
The modified data frame with the added gt column.
}
\description{
Compute the generation time values for the transmission pairs.
}
