% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups_utils.R
\name{get_peak}
\alias{get_peak}
\title{Get the estimated peak values for each group in the data.}
\usage{
get_peak(data)
}
\arguments{
\item{data}{A data frame containing the necessary columns "group" and "date_onset".}
}
\value{
A named list of estimated peak values for each group.
}
\description{
This function calculates the estimated peak values for each group in the provided data. The data should have columns named "group" and "date_onset".
}
\seealso{
\code{\link[incidence:incidence]{incidence::incidence()}} \code{\link[incidence:estimate_peak]{incidence::estimate_peak()}}
}
\keyword{internal}
