% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_susceptibles.R
\name{get_prop_susceptibles}
\alias{get_prop_susceptibles}
\title{Get Proportion of Susceptibles within a Time Window}
\usage{
get_prop_susceptibles(data, t, n_groups, size, name)
}
\arguments{
\item{data}{A linelist.}

\item{t}{The time point (inclusive).}

\item{n_groups}{The total number of groups.}

\item{size}{A numeric vector of length \code{n_groups} containing the initial known sizes of the groups.}

\item{name}{A character vector of length \code{n_groups} containing the names of the groups.}
}
\value{
A numeric vector with the calculated proportions of susceptibles in each group.
}
\description{
Calculates the proportion of susceptibles in each group within a specified time window.
}
\examples{
data <- data.frame(
  group = c("A", "B", "B", "A"),
  id = c("HmPsw2", "WtYSxS", "2Kxtgd", "xtgn1T"),
  source = c(NA, NA, "WtYSxS", "WtYSxS"),
  source_group = c(NA, NA, "B", "B"),
  date_onset = c(0, 0, 2, 3)
)

n_groups <- 2
size <- c(10, 15)
name <- c("A", "B")
t <- 2

get_prop_susceptibles(data, t, n_groups, size, name)
}
