% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups_utils.R
\name{get_Ri}
\alias{get_Ri}
\title{Returns the Case Reproduction Number}
\usage{
get_Ri(data)
}
\arguments{
\item{data}{A data frame with columns for source ID and group.}
}
\value{
The data frame with the number of cases each ID generated in each group.
}
\description{
This function takes in a data frame with columns for source ID and group, and returns the data frame with the number of cases each ID generated in each group.
}
\examples{
data <- data.frame(id = c(1, 2, 3, 4, 5),
                   source = c(NA, 1, 2, 2, 3),
                   group = c("A", "A", "B", "B", "A"))
get_Ri(data)
}
