% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups_utils.R
\name{get_si}
\alias{get_si}
\title{Compute serial interval}
\usage{
get_si(data)
}
\arguments{
\item{data}{The data element of \code{simulate_groups}. Please, name \code{date_onset} for the date of symptom onset column.}
}
\value{
The modified data frame with the added si column.
}
\description{
Compute the serial interval values for the transmission pairs.
}
