% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_delta_optim.R
\name{M0_to_delta_optim}
\alias{M0_to_delta_optim}
\title{Retrieve the assortativity coefficient from the next generation matrix 'M0'}
\usage{
M0_to_delta_optim(n_groups, name, size, r0, M0_observed)
}
\arguments{
\item{n_groups}{Number of groups}

\item{name}{A character vector of length \code{n_groups} containing the names of the groups (default is A,B,C...)}

\item{size}{A numeric vector of length \code{n_groups} containing the sizes of the groups}

\item{r0}{A numeric vector of length \code{n_groups} containing the basic Reproduction Number (R0) of the groups}

\item{M0_observed}{The next generation matrix}
}
\value{
A vector with the delta coefficients for each group
}
\description{
The next generation matrix is a matrix representing the number of secondary
cases generated by an infector from a particular group (column) in a
totally susceptible population in each other group (row).
}
\details{
M0_to_delta_optim can retrieve the assortativity coeffcient \code{delta}
in the early phase of the epidemic.
}
\examples{


M0_observed <- generate_M0(
  n_groups = 2,
  size = c(50, 100),
  name = c("A", "B"),
  r0 = c(5, 3),
  delta = c(2, 5.3))

M0_to_delta_optim(
  n_groups = 2,
  size = c(50, 100),
  name = c("A", "B"),
  r0 = c(5, 3),
  M0_observed = M0_observed)

}
\seealso{
\code{\link{generate_M0}}
}
