% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_to_delta.R
\name{Mcol_to_delta}
\alias{Mcol_to_delta}
\title{Retrieve the assortativity coefficient from the mixing matrix 'Mcol'}
\usage{
Mcol_to_delta(n_groups, name = NULL, size, Mcol_observed)
}
\arguments{
\item{n_groups}{Number of groups}

\item{name}{Optional. A character vector of length \code{n_groups} containing the names of the groups (default is A,B,C...)}

\item{size}{A numeric vector of length \code{n_groups} containing the sizes of the groups}

\item{Mcol_observed}{The mixing matrix}
}
\value{
A vector with the delta coefficients for each group
}
\description{
The mixing matrix Mcol is a matrix representing the proportions of
cases generated by an infector from a particular group (column) in a
totally susceptible population in each other group (row).
}
\details{
Mcol_to_delta can retrieve the assortativity coefficient \code{delta}
in the early phase of the epidemic.
}
\examples{
n_groups = 4
name =LETTERS[1:n_groups]
size = round(runif(n_groups, min = 10, max = 1e3))
delta = runif(n_groups, min = 0.01, max = 1e3)

Mcol_observed <- generate_Mcol(n_groups, name, size, delta)
Mcol_to_delta( n_groups = n_groups, size = size, Mcol_observed = Mcol_observed)

}
\seealso{
\code{\link{generate_Mcol}}
}
