% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_delta_optim.R
\name{Mcol_to_delta_optim}
\alias{Mcol_to_delta_optim}
\title{Retrieve the assortativity coefficient from the mixing matrix 'Mcol'}
\usage{
Mcol_to_delta_optim(n_groups, name, size, Mcol_observed)
}
\arguments{
\item{n_groups}{Number of groups}

\item{name}{A character vector of length \code{n_groups} containing the names of the groups (default is A,B,C...)}

\item{size}{A numeric vector of length \code{n_groups} containing the sizes of the groups}

\item{Mcol_observed}{The mixing matrix}
}
\value{
A vector with the delta coefficients for each group
}
\description{
The mixing matrix Mcol is a matrix representing the proportions of
cases generated by an infector from a particular group (column) in a
totally susceptible population in each other group (row).
}
\details{
Mcol_to_delta_optim can retrieve the assortativity coefficient \code{delta}
in the early phase of the epidemic.
}
\examples{


Mcol_observed <- generate_Mcol(
  n_groups = 2,
  size = c(50, 100),
  name = c("A", "B"),
  delta = c(2, 5.3))

Mcol_to_delta_optim(
  n_groups = 2,
  size = c(50, 100),
  name = c("A", "B"),
  Mcol_observed = Mcol_observed)

}
\seealso{
\code{\link{generate_Mcol}}
}
