% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_delta_optim.R
\name{Mcolt_to_delta_optim}
\alias{Mcolt_to_delta_optim}
\title{Retrieve the assortativity coefficient from the effective mixing matrix 'Mcolt'}
\usage{
Mcolt_to_delta_optim(
  n_groups,
  name,
  size,
  r0,
  prop_susceptible,
  Mcolt_observed
)
}
\arguments{
\item{n_groups}{Number of groups}

\item{name}{A character vector of length \code{n_groups} containing the names of the groups}

\item{size}{A numeric vector of length \code{n_groups} containing the sizes of the groups}

\item{r0}{A numeric vector of length \code{n_groups} containing the basic Reproduction Number (R0) of the groups}

\item{prop_susceptible}{A numeric vector of length \code{n_groups} containing the proportion of susceptible in each group at time t}

\item{Mcolt_observed}{The effective next generation matrix}
}
\value{
A vector with the delta coefficients for each group
}
\description{
The effective next generation matrix refers to a matrix representing the
number of secondary cases generated by an infector from a particular group (column)
in each other group (row) at time t.
}
\details{
Mcolt_to_delta_optim can retrieve the assortativity coefficient \code{delta}
at any phase of the epidemic.
}
\examples{

n_groups = 2
size = c(1e4, 1e5)
name = LETTERS[1:n_groups]
r0 = c(5, 3)
prop_susceptible = c(0.6, 0.3)
delta = c(10, 2)

Mcolt_observed <- generate_Mcolt(
  n_groups = n_groups,
  size = size,
  name = name,
  r0 = r0,
  prop_susceptible = prop_susceptible,
  delta = delta
)

Mcolt_to_delta_optim(
  n_groups = n_groups,
  size = size,
  name = name,
  r0 = r0,
  prop_susceptible = prop_susceptible,
  Mcolt_observed = Mcolt_observed
)

}
\seealso{
\code{\link{generate_Mcol}}
}
