% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mix_window.R
\name{mix_window_data}
\alias{mix_window_data}
\title{Calculate Mixing Window Data}
\usage{
mix_window_data(data, window_length)
}
\arguments{
\item{data}{A data frame containing the data for calculation.}

\item{window_length}{The length of each mixing window.}
}
\value{
A data frame containing the mixing window information.
}
\description{
This function calculates the mixing window data based on the provided data and window length. It returns a data frame containing the mixing window information, including frequencies, confidence intervals, and window boundaries.
}
\keyword{internal}
