% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_to_delta.R
\name{Mt_to_delta}
\alias{Mt_to_delta}
\title{Calculate Assortativity Coefficients from Observed Mt Matrix}
\usage{
Mt_to_delta(n_groups, name = NULL, size, r0, prop_susceptible, Mt_observed)
}
\arguments{
\item{n_groups}{Number of groups}

\item{name}{Optional. A character vector of length \code{n_groups} containing the names of the groups.}

\item{size}{A numeric vector of length \code{n_groups} containing the sizes of the groups}

\item{r0}{A numeric vector of length \code{n_groups} containing the basic reproductive number of the disease in each group}

\item{prop_susceptible}{A numeric vector of length \code{n_groups} containing the proportion of susceptibles in each group (between 0 and 1)}

\item{Mt_observed}{The observed Mt matrix}
}
\value{
A vector with the calculated assortativity coefficients (\code{delta}) for each group
}
\description{
The Mt_to_delta function calculates the assortativity coefficients (\code{delta}) based on the observed matrix \code{Mt}.
The matrix \code{Mt} represents the number of secondary cases generated by an infector from a particular group (column) in each other group (row) at a specific time \code{t}.
}
\examples{
n_groups = 2
name = LETTERS[1:n_groups]
size = c(100, 100)
delta = c(11, 2.5)
r0 = c(4, 3)
prop_susceptible = c(0.1, 0.7)
Mt_observed <-
 generate_Mt(n_groups, name, size, r0, prop_susceptible, delta)
Mt_to_delta(
 n_groups = n_groups,
 name = name,
 size = size,
 r0 = r0,
 prop_susceptible = prop_susceptible,
 Mt_observed = Mt_observed
)

}
