% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mix_window.R
\name{plot_mix_window}
\alias{plot_mix_window}
\title{Plot Mixing Windows}
\usage{
plot_mix_window(data, window_length, n_groups, size, name, delta)
}
\arguments{
\item{data}{A data frame containing the data for calculation.}

\item{window_length}{The length of each mixing window.}
}
\value{
A plot showing the mixing windows with confidence intervals and ground truth values.
}
\description{
This function calculates the mixing window data based on the provided data and window length, and then plots the mixing windows. It returns a plot showing the mixing windows for each group and source group, along with confidence intervals and a dotted line representing the ground truth values.
}
\keyword{internal}
