% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Ri.R
\name{plot_Ri}
\alias{plot_Ri}
\title{Plot Ri Values}
\usage{
plot_Ri(data, name, r0)
}
\arguments{
\item{data}{A data frame containing the data for calculation.}

\item{name}{The name of the variable to be used for Ri calculation.}

\item{r0}{The basic reproduction numbers to plot the hlines.}
}
\value{
A combined plot of Ri values for facet and group views.
}
\description{
This function calculates Ri values based on the provided data and name, and then plots the Ri values. It returns a combined plot with facet and group views.
}
