% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Ri.R
\name{plot_Ri_facet}
\alias{plot_Ri_facet}
\title{Plot Ri Values for Facet View}
\usage{
plot_Ri_facet(Ri_facet)
}
\arguments{
\item{Ri_facet}{A data frame containing Ri values for each facet and group.}
}
\value{
A plot of Ri values for the facet view.
}
\description{
This function plots Ri values for the facet view based on the provided Ri_facet data frame.
}
\keyword{internal}
