% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Ri.R
\name{plot_Ri_group}
\alias{plot_Ri_group}
\title{Plot Ri Values for Group View}
\usage{
plot_Ri_group(Ri_group, name = name, r0 = r0)
}
\arguments{
\item{Ri_group}{A data frame containing Ri values for each group.}

\item{name}{The name of the variable to be used for Ri calculation.}

\item{r0}{The basic reproduction numbers to plot the hlines.}
}
\value{
A plot of Ri values for the group view.
}
\description{
This function plots Ri values for the group view based on the provided Ri_group data frame.
}
\keyword{internal}
