% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stats.R
\name{plot_stats}
\alias{plot_stats}
\title{Plot incidence and proportion of susceptible over time}
\usage{
plot_stats(stats)
}
\arguments{
\item{stats}{A dataframe returned by simulate_groups()$stats}
}
\value{
A list containing two ggplots: incidence and proportion of susceptible over time.
}
\description{
This function generates two plots: incidence and proportion of susceptible over time.
}
