% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{plot_tree}
\alias{plot_tree}
\title{Plots the transmission tree over time}
\usage{
plot_tree(data, pal)
}
\arguments{
\item{data}{A dataframe from simulate_groups().}

\item{pal}{A named character vector of hex codes for each group.}
}
\description{
Plots the transmission tree over time
}
\examples{
out <-
simulate_groups(
 duration = 100,
 n_groups = 2,
 size = c(10, 30),
 name = c("A", "B"),
 delta = c(5, 2),
 intro_n = c(1, 2),
 r0 = c(4, 3),
 generation_time = simulacr::make_disc_gamma(mean = 5, sd = 2)$d(1:30)
)
plot_tree(out$data, pal = c("A" = "#FA05BA", "B" = "#FFA500"))
}
