% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_simulation.R
\name{process_simulation}
\alias{process_simulation}
\title{Process Simulation Results}
\usage{
process_simulation(simulation_data, peak_coeff, size, name)
}
\arguments{
\item{simulation_data}{A data frame containing the simulation results, with columns for simulation ID, group, and other relevant variables.}

\item{peak_coeff}{A numeric value representing the peak coefficient for calculating the maximum time range for delta estimation.}

\item{size}{A vector of integers representing the sizes of each group.}

\item{name}{A vector of characters representing the names of each group.}
}
\value{
A data frame containing the delta estimates for each group and simulation.
}
\description{
This function processes the results of a simulation by calculating delta estimates for each group and simulation based on the provided peak coefficient.
}
\keyword{internal}
