% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups_utils.R
\name{reverse_scale}
\alias{reverse_scale}
\title{Reverse the standardisation of delta.}
\usage{
reverse_scale(scaled_delta)
}
\arguments{
\item{scaled_delta}{A scaled numeric value to be reversed.}
}
\value{
The original unscaled delta value.
}
\description{
This function takes a scaled delta value 'scaled_delta' and reverses
the standardisation by applying the formula: (1 + scaled_delta) / (1 - scaled_delta).
}
\details{
\code{scaled_delta} is a real number constrained within the interval \link{-1, 1}.
\code{scaled_delta} < 0 indicates dissortativity.
\code{scaled_delta} > 0 indicates assortativity.
\code{scaled_delta} = 0 indicates neutrality.

\code{raw_delta} is a non-negative real number constrained within the interval [0, ∞).
\code{raw_delta} < 1 indicates dissortativity.
\code{raw_delta} > 1 indicates assortativity.
\code{raw_delta} = 1 indicates neutrality.
}
