% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Ri.R
\name{Ri_data}
\alias{Ri_data}
\title{Calculate Ri data for each date of onset}
\usage{
Ri_data(data, name)
}
\arguments{
\item{data}{A data frame containing the data for calculation.}

\item{name}{The name of the variable to be used for Ri calculation.}
}
\value{
A list with two data frames: Ri_facet and Ri_group.
}
\description{
This function calculates Ri values based on the provided data and name. It returns a list containing the Ri values for each facet and group.
}
\keyword{internal}
