% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups_utils.R
\name{sample_sources}
\alias{sample_sources}
\title{Samples infectors based on their Force of Infection}
\usage{
sample_sources(group, n_cases, data, indivFOI)
}
\arguments{
\item{group}{The current group of interest}

\item{n_cases}{The number of new cases to be generated in that group at time t}

\item{data}{The data/linelist at time t}

\item{indivFOI}{The individual FOI for all potential infectors targeting the group of interest (will be used as weight for sampling)}
}
\value{
A list of source IDs
}
\description{
This functions samples infector IDs based on their FOI targeting the group of interest
}
\keyword{internal}
