% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups_utils.R
\name{scale}
\alias{scale}
\title{Scale raw delta values.}
\usage{
scale(raw_delta)
}
\arguments{
\item{raw_delta}{A numeric value to be scaled.}
}
\value{
The scaled delta value.
}
\description{
This function takes a numeric value 'raw_delta' and scales it by
applying the formula: (raw_delta - 1) / (raw_delta + 1) for finite values,
and returns 1.0 for Infinite values.
}
\details{
\code{raw_delta} is a non-negative real number constrained within the interval [0, ∞).
\code{raw_delta} < 1 indicates dissortativity.
\code{raw_delta} > 1 indicates assortativity.
\code{raw_delta} = 1 indicates neutrality.

\code{scaled_delta} is a real number constrained within the interval \link{-1, 1}.
\code{scaled_delta} < 0 indicates dissortativity.
\code{scaled_delta} > 0 indicates assortativity.
\code{scaled_delta} = 0 indicates neutrality.
}
