% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups.R
\name{simulate_groups}
\alias{simulate_groups}
\title{simulate_groups}
\usage{
simulate_groups(
  duration = 100,
  n_groups,
  size,
  name = NULL,
  delta = rep(1, n_groups),
  intro_n = NULL,
  r0,
  generation_time,
  incubation_period = NULL,
  dt = 1,
  quietly = TRUE
)
}
\arguments{
\item{duration}{An integer value representing the duration of the simulation}

\item{n_groups}{An integer value representing the number of groups}

\item{size}{A vector of integers representing the sizes of each group}

\item{name}{A vector of characters representing the names of each group}

\item{delta}{A numeric vector representing the within-group transmission factor (assortativity coefficient) for each group}

\item{intro_n}{A vector of n introductions respective to each group.}

\item{r0}{A numeric vector representing the basic reproductive number (R0) for each group}

\item{generation_time}{The probability mass function (pmf) of the generation time (sums to 1).}

\item{incubation_period}{Optional. A vector of integers that will be sampled with replacement.}

\item{dt}{The time step of the simulation (default = 1 day)}

\item{quietly}{Logical. Whether to print progress or not. Default is TRUE (doesn't print progress).}
}
\value{
A list containing the individual linelist, the summary statistics per time step & input matrices (Mcol/M0).
}
\description{
Simulate the spread of an infectious disease among multiple groups
}
\examples{
simulate_groups(
  duration = 100,
  n_groups = 3,
  size = c(100, 50, 200),
  name = c("A", "B", "C"),
  delta = c(1, 2, 5),
  intro_n = c(0, 7, 5),
  r0 = c(2, 3, 4),
  generation_time = c(0.1, 0.3, 0.4, 0.2))


}
