% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_M.R
\name{generate_M0}
\alias{generate_M0}
\title{Generate M0 Matrix}
\usage{
generate_M0(group_n, size, gamma = rep(1, group_n), r0, name = NULL)
}
\arguments{
\item{group_n}{An integer specifying the number of groups.}

\item{size}{A numeric vector specifying the size of each group.}

\item{gamma}{A numeric vector specifying the gamma values for each group. Default is a vector of 1s.}

\item{r0}{A numeric vector specifying the basic reproduction number (R0) for each group.}

\item{name}{A character vector specifying the names of the groups. Default is NULL, in which case the groups are named using the first 'group_n' letters of the alphabet.}
}
\value{
Basic Reproduction Number Matrix, rows are 'from' groups and columns are 'to' groups.
}
\description{
Basic Reproduction Number Matrix. Each row represent 'from' groups and columns represent 'to' groups.
Conditional on an individual in group 'i' seeding infection, the expected number of secondary infections within group 'j' is given by M0\link{i,j}. The sum of each row is the basic reproduction number (R0) for each group.
}
\keyword{internal}
