% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_M.R
\name{generate_Mcol}
\alias{generate_Mcol}
\title{Generate Mcol matrix}
\usage{
generate_Mcol(n_groups, name = NULL, size, delta = rep(1, n_groups))
}
\arguments{
\item{n_groups}{Number of groups}

\item{name}{A character vector of length \code{n_groups} containing the names of the groups (default is A,B,C...)}

\item{size}{A numeric vector of length \code{n_groups} containing the sizes of the groups}

\item{delta}{A numeric vector of length \code{n_groups} containing the assortativity coefficient for each group (default is 1 for all groups)}
}
\value{
A matrix representing the number of secondary cases generated by an infector from a particular group (column) in a totally susceptible population in each other group (row), taking into the receiving group size and the assortativity coefficient.
}
\description{
Generates a matrix representing the probability of an infector from a particular group (column) transmitting the infection to every group (row). It accounts for the receiving group size & the assortativity coefficient (delta).
}
\examples{
generate_Mcol(
  n_groups = 2,
  name = c("A", "B"),
  size = c(100, 100),
  delta = c(2, 2)
)

}
