% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups_utils.R
\name{simulate_groups_furrr}
\alias{simulate_groups_furrr}
\title{simulate_groups_furrr}
\usage{
simulate_groups_furrr(
  n_simulations,
  n_cores = as.integer(future::availableCores() - 1),
  duration,
  n_groups,
  size,
  name,
  delta,
  intro_n,
  r0,
  generation_time,
  incubation_period = NULL,
  dt = 1,
  stack = TRUE
)
}
\arguments{
\item{n_simulations}{The number of simulations to run.}

\item{n_cores}{The number of cores to use in parallel. Default is as.integer(future::availableCores() - 1).}

\item{duration}{An integer value representing the duration of the simulation.}

\item{n_groups}{An integer value representing the number of groups.}

\item{size}{A vector of integers representing the sizes of each group.}

\item{name}{A vector of characters representing the names of each group.}

\item{delta}{A numeric vector representing the within-group transmission factor (assortativity coefficient) for each group.}

\item{intro_n}{A vector of integers representing the number of introductions for each group.}

\item{r0}{A numeric vector representing the basic reproductive number (R0) for each group.}

\item{generation_time}{The probability mass function (pmf) of the generation time.}

\item{incubation_period}{Optional. A vector of integers that will be sampled with replacement.}

\item{dt}{The time step of the simulation (default = 1 day).}

\item{stack}{Logical. If TRUE, returns a stacked dataframe by simulation ID else returns a list of \code{simulate_groups()} outputs. Default is TRUE.}
}
\value{
A list containing data frames with information about the simulated groups and the statistics of the simulation & matrices inputs (Mcol/M0).
}
\description{
Simulate multiple outbreaks in parallel. A furrr wrapper for \code{simulate_groups}.
}
\examples{
\dontrun{
library(furrr)
simulate_groups_furrr(
  n_simulations = 20,
  n_cores = 5,
  duration = 100,
  n_groups = 3,
  size = c(100, 50, 200),
  name = c("A", "B", "C"),
  delta = c(1, 2, 5),
  intro_n = c(7, 5),
  r0 = c(2, 3, 4),
  generation_time = c(0.1, 0.3, 0.4, 0.2))
}


}
